requirements.txt
torch>=2.0.0
torchvision>=0.24.0
transformers>=4.30.0
datasets>=2.0.0
matplotlib>=3.5.0
Pillow>=9.0.0
